Add refresh of attester when modified and once a day
This allows new attesters to be added without a new installation/upgrade
of the extension.
