+++ platform/firefox/index.ts
@@ -3,12 +3,15 @@ import {
     handleBeforeSendHeaders,
     handleHeadersReceived,
     handleInstall,
+    handleStartup,
 } from '../../src/background';
 
 const BROWSER = BROWSERS.FIREFOX;
 const STORAGE = browser.storage.local;
 
 chrome.runtime.onInstalled.addListener(handleInstall(STORAGE));
+
+chrome.runtime.onStartup.addListener(handleStartup(STORAGE));
 
 chrome.webRequest.onBeforeSendHeaders.addListener(
     handleBeforeSendHeaders(STORAGE),
