+++ src/common/const.ts
@@ -1,0 +1,1 @@
+export const STORAGE_ID_ATTESTER_CONFIGURATION = 'attesters_by_issuer_key';
