+++ src/common/settings.ts
@@ -1,3 +1,7 @@
+import { IssuerConfig } from '@cloudflare/privacypass-ts';
+import { STORAGE_ID_ATTESTER_CONFIGURATION } from './const';
+import { getLogger } from './logger';
+
 export const SETTING_NAMES = {
     SERVICE_WORKER_MODE: 'serviceWorkerMode',
     ATTESTERS: 'attesters',
@@ -54,7 +58,44 @@ export function getSettings(): Settings {
     return settings;
 }
 
-export function saveSettings(
+export const refreshAttesterLookupByIssuerKey = async (storage: chrome.storage.StorageArea) => {
+    // Force reset associations between public keys and issuers that trust each attester that we know about
+    const attestersByIssuerKey = new Map<string, string>();
+
+    const { attesters, serviceWorkerMode: mode } = getSettings();
+    const logger = getLogger(mode);
+
+    // Populate issuer keys for issuers that trust our ATTESTERS
+    for (const attester of attesters) {
+        const nowTimestamp = Date.now();
+
+        // Connect to each of the ATTESTERS we know about for their issuer directory
+        const response = await fetch(`${attester}/v1/private-token-issuer-directory`);
+        if (!response.ok) {
+            logger.log(`"${attester}" issuer keys not available, attester will not be used.`);
+            return;
+        }
+
+        const directory: Record<string, IssuerConfig> = await response.json();
+        // for each attester in the directory
+        for (const { 'token-keys': tokenKeys } of Object.values(directory)) {
+            for (const key of tokenKeys) {
+                const notBefore = key['not-before'];
+                if (!notBefore || notBefore > nowTimestamp) {
+                    continue;
+                }
+                attestersByIssuerKey.set(key['token-key'], attester);
+            }
+        }
+        storage.set({
+            [STORAGE_ID_ATTESTER_CONFIGURATION]: Object.fromEntries(attestersByIssuerKey),
+        });
+    }
+
+    logger.info('Attester lookup by Issuer key populated');
+};
+
+export async function saveSettings(
     storage: chrome.storage.StorageArea,
     name: SettingsKeys,
     value: string,
@@ -62,6 +103,7 @@ export function saveSettings(
     switch (name) {
         case 'attesters':
             settings[name] = rawSettingToSettingAttester(value);
+            await refreshAttesterLookupByIssuerKey(storage);
             break;
         case 'serviceWorkerMode':
             if (
